cs50p/project/requirements.txt
selenium
schedule
yagmail
python-dotenv
pytest
pytest-mock
